NullableDateTimeHandler: extract more coverage by exploiting parent's method
